cleanup and docs
